requirements.txt
SpeechRecognition==3.10.0
pyttsx3==2.90
PyAudio==0.2.14
python-dotenv==1.0.0
SQLAlchemy==2.0.23
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
openai>=1.6.1
tenacity==8.2.3 